enhance/90: moved responce messges in constants file
